requirements.txt
alabaster==0.7.13
asgiref==3.7.2
Babel==2.12.1
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
Django==4.2.4
django-address==0.2.8
django-dirtyfields==1.9.2
docutils==0.20.1
furo==2023.8.19
griffe==0.36.0
idna==3.4
imagesize==1.4.1
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
Pygments==2.16.1
requests==2.31.0
snowballstemmer==2.2.0
soupsieve==2.4.1
Sphinx==7.2.5
sphinx-autodoc-typehints==1.24.0
sphinx-basic-ng==1.0.0b2
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
